Respond to transfer bids that arent winning bids

+++ app/models/transfer.rb
@@ -41,21 +41,23 @@ class Transfer < ActiveRecord::Base
   end
 
   def complete_transfer
-    self.get_team_player_involved.active_transfers.where.not(id: self.id).each do |losing_transfer|
-      losing_transfer.destroy
-    end
-    self.transfer_items.each do |transfer_item|
-      receiver = transfer_item.receiving_team
-      sender = transfer_item.sending_team
-      if transfer_item.transfer_item_type === "Cash"
-        sender.decrement!(:cash_balance_cents, transfer_item.cash_cents)
-        receiver.increment!(:cash_balance_cents, transfer_item.cash_cents)
-        receiver.delist_involuntarily_listed_team_players if receiver.should_be_back_in_the_black
-      elsif transfer_item.transfer_item_type === "Player"
-        contract = transfer_item.team_player.current_contract
-        contract.update_attributes!(signed: false)
-        contract.update_attributes!(team: receiver, signed: true)
-        transfer_item.team_player.update_attributes!(team: receiver, first_team: false, squad_position: SquadPosition.find_by(short_name: 'SUB'), transfer_minimum_bid: nil, transfer_completes_at: nil, is_voluntary_transfer: false)
+    if transfer_completed?
+      self.get_team_player_involved.active_transfers.where.not(id: self.id).each do |losing_transfer|
+        losing_transfer.destroy
+      end
+      self.transfer_items.each do |transfer_item|
+        receiver = transfer_item.receiving_team
+        sender = transfer_item.sending_team
+        if transfer_item.transfer_item_type === "Cash"
+          sender.decrement!(:cash_balance_cents, transfer_item.cash_cents)
+          receiver.increment!(:cash_balance_cents, transfer_item.cash_cents)
+          receiver.delist_involuntarily_listed_team_players if receiver.should_be_back_in_the_black
+        elsif transfer_item.transfer_item_type === "Player"
+          contract = transfer_item.team_player.current_contract
+          contract.update_attributes!(signed: false, ends_at: 90.days.from_now.strftime('%Y-%m-%d'))
+          contract.update_attributes!(team: receiver, signed: true)
+          transfer_item.team_player.update_attributes!(team: receiver, first_team: false, squad_position: SquadPosition.find_by(short_name: 'SUB'), transfer_minimum_bid: nil, transfer_completes_at: nil, is_voluntary_transfer: false)
+        end
       end
     end
   end
@@ -98,6 +100,13 @@ class Transfer < ActiveRecord::Base
 
   def get_cash_involved
     get_cash_transfer_item.cash_cents
+  end
+
+  def make_counter_offer(team, counter_offer)
+    if team_transfer_status(team) != 'accepted'
+      get_cash_transfer_item.update_attributes(cash_cents: counter_offer)
+      self.update_attributes(primary_team_accepted: self.primary_team == team, secondary_team_accepted: self.secondary_team == team)
+    end
   end
 
  	private
